+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Unreleased
 ----------
 - Added `ApiInfo::new` constructor
+- Bumped `http-endpoint` dependency to `0.3`
 - Bumped `websocket-util` dependency to `0.5`
 - Bumped `async-tungstenite` dependency to `0.5`